requirements.txt
blinker==1.9.0
bokeh==3.7.2
click==8.1.8
contourpy==1.3.1
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==8.0.15
narwhals==1.32.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==6.30.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
tornado==6.4.2
tzdata==2025.2
Werkzeug==3.1.3
xyzservices==2025.1.0
